cuda: gate_air 19-col AIR kernel + composition allocation overflow fix
Update the gate_air constraint kernel for the trace-shrink AIR (19 main / 20
interaction cols): masks, rc single-d lookup, ts-range reconstruction, tile
offsets. Widen the uint32 allocator helpers (cuda_malloc_uint32_t /
cuda_alloc_zeroes_uint32_t) and their FFI bindings from int to size_t, and
correct the composition `numerators` buffer to a u32 count (4*eval_domain_size,
was 16*): at 2^26 the old int arg truncated 2^31 to INT_MIN and requested a
garbage size. Byte-identical to the resident path.

+++ crates/stwo/src/prover/backend/cuda/column.rs
@@ -111,7 +111,7 @@ impl Column<BaseField> for interface::base_field_vec::BaseFieldVec {
 
     unsafe fn uninitialized(len: usize) -> Self {
         Self {
-            device_ptr: bindings::cuda_malloc_uint32_t(len as u32),
+            device_ptr: bindings::cuda_malloc_uint32_t(len),
             size: len,
             owns_memory: true,
         }
@@ -192,7 +192,7 @@ impl Column<SecureField> for SecureFieldVec {
 
     unsafe fn uninitialized(len: usize) -> Self {
         Self {
-            device_ptr: bindings::cuda_malloc_uint32_t(4 * len as u32),
+            device_ptr: bindings::cuda_malloc_uint32_t(4 * len),
             size: len,
         }
     }
