+++ crates/stwo/src/stwo_cuda/base_field_vec.rs
@@ -44,12 +44,12 @@ impl BaseFieldVec {
     }
 
     pub fn new_uninitialized(size: usize) -> Self {
-        let device_ptr = unsafe { bindings::cuda_malloc_uint32_t(size as u32) };
+        let device_ptr = unsafe { bindings::cuda_malloc_uint32_t(size) };
         Self::new(device_ptr, size)
     }
 
     pub fn new_zeroes(size: usize) -> Self {
-        let device_ptr = unsafe { bindings::cuda_alloc_zeroes_uint32_t(size as u32) };
+        let device_ptr = unsafe { bindings::cuda_alloc_zeroes_uint32_t(size) };
         Self::new(device_ptr, size)
     }
 
@@ -315,12 +315,12 @@ impl Uint32Vec {
     }
 
     pub fn new_uninitialized(size: usize) -> Self {
-        Self::new(unsafe { bindings::cuda_malloc_uint32_t(size as u32) }, size)
+        Self::new(unsafe { bindings::cuda_malloc_uint32_t(size) }, size)
     }
 
     pub fn new_zeroes(size: usize) -> Self {
         Self::new(
-            unsafe { bindings::cuda_alloc_zeroes_uint32_t(size as u32) },
+            unsafe { bindings::cuda_alloc_zeroes_uint32_t(size) },
             size,
         )
     }
@@ -483,14 +483,14 @@ impl Uint128Vec {
 
     pub fn new_uninitialized(size: usize) -> Self {
         Self::new(
-            unsafe { bindings::cuda_malloc_uint32_t(4 * size as u32) },
+            unsafe { bindings::cuda_malloc_uint32_t(4 * size) },
             size,
         )
     }
 
     pub fn new_zeroes(size: usize) -> Self {
         Self::new(
-            unsafe { bindings::cuda_alloc_zeroes_uint32_t(4 * size as u32) },
+            unsafe { bindings::cuda_alloc_zeroes_uint32_t(4 * size) },
             size,
         )
     }
